Move ExperimentHeader to activity folder
