+++ ui/src/components/activity/ActivityHeader/MarkAsCompletedButton/index.js
